Finished making the views.
This does not yet include the styles file or the controller to actually make it all work. Additionally, I discovered that the embedding of waves will be a little more complicated than initially thought. Multiple branches will be made to determine the best solution to the problem (see the home.haml view for a full explanation).